+ initial commit, imported from my ancient subversion tilde bins
this was in my 2006 ~/bin/ svn repo, it hasn't changed since 2008, the
log reads that it was imported from pre-existing scripts (prior to any
versioning), so this probably existed since at least early 2000s.
there's only one commit since svn import, to add support for bulleted
lists that cover more than one line.  since there's little change, no
need to retain the svn history.

this includes the whole kit: template, makefile, and the awk and sed
scripts, plus we made a README since it'll be on github.

TODO: should make a wrapper that copies the template and replaces some
words.  also need to make a local branch that has my name and address in
it rather than the dummy from template, because it will get annoying to
have to edit that every time