/ header: m4: filename macro needs no space after or syntax error
for some reason this worked fine on my debian system but has an issue on
my ubuntu system when there's a space, it can't find the filename and
thinks the space is part thereof.  removing the trailing space did the
trick.  not sure if this is wrong or not, but it works on both systems
this way